CBL-7240 : Add mdns logging domain (Port) (#3453)
* Port changes from release/3.3 branch for CBL-7394.

* Added mdns logging domain to both objective-c and swift

* Reserved 128 and 256 value for bluetooth and wifi-aware that we may have in the future for multipeer replicator.

* Changed Swift LogDomain type from UInt8 to UInt.

* Fixed missing switch cases in CBLFileLogSink.mm.

* Skip testConfigurationValidation as precondition cannot be asserted

* We use Swift’s precondition() for parameter validation in Precondition.swift. Since it traps like fatalError, it cannot be asserted in tests, so the test is left disabled for manual runs.

* Removed leftover debug logging

* (New Change) : Update LogDomains’s rawValue types to UInt to match LogDomain’s rawValue type. It doesn’t matter much but it’s better that the type to be in sync so no type conversion needed.

+++ Swift/Tests/LogSinkTest.swift
@@ -376,7 +376,7 @@ class LogSinkTest: CBLTestCase {
             let logSink = TestCustomLogSink()
             LogSinks.custom = CustomLogSink(level: .debug, domains: domains[i], logSink: logSink)
             for j in 0..<domains.count {
-                Log.log(domain: LogDomain(rawValue: UInt8(domains[j].rawValue))!, level: .verbose, message: names[j])
+                Log.log(domain: LogDomain(rawValue: domains[j].rawValue)!, level: .verbose, message: names[j])
             }
             XCTAssertEqual(logSink.lines.count, 1)
             XCTAssertEqual(logSink.lines[0], names[i])
@@ -392,7 +392,7 @@ class LogSinkTest: CBLTestCase {
             let logSink = TestCustomLogSink()
             LogSinks.custom = CustomLogSink(level: .debug, domains: combined, logSink: logSink)
             for j in 0..<domains.count {
-                Log.log(domain: LogDomain(rawValue: UInt8(domains[j].rawValue))!, level: .verbose, message: names[j])
+                Log.log(domain: LogDomain(rawValue: domains[j].rawValue)!, level: .verbose, message: names[j])
             }
             
             XCTAssertEqual(logSink.lines.count, i + 1)
@@ -405,7 +405,7 @@ class LogSinkTest: CBLTestCase {
         let logSink = TestCustomLogSink()
         LogSinks.custom = CustomLogSink(level: .debug, domains: .all, logSink: logSink)
         for i in 0..<domains.count {
-            Log.log(domain: LogDomain(rawValue: UInt8(domains[i].rawValue))!, level: .verbose, message: names[i])
+            Log.log(domain: LogDomain(rawValue: domains[i].rawValue)!, level: .verbose, message: names[i])
         }
         XCTAssertEqual(logSink.lines.count, names.count)
         for i in 0..<names.count {
